Implementing Distribution Fitting Pipeline
Allows for fitting of distributions to the processed coverage data.

Given a processed coverage file, fitDistribution.R fits a gamma
distribution to the data. It then can save the found parameters,
as well as find and save a p-value for every bin.

At the moment, bin p-values are stored in a separate file from
the rest of the data. This will be adjusted in future commits.

Also contained is a first pass version of runEntirePipeline.sh,
to be renamed and adjusted to work as expected from a Docker
script. For now, it works from start to finish to generate
a coverage file, a processed coverage file, a file of p-values,
anda file containing the fitted gamma distribution parameters.